Generate Raw STUDCOMP Data

Generate Raw STUDCOMP based on STUDCOMP.yaml from gsm.mapping.

Usage

Raw_STUDCOMP(data, previous_data, spec, ...)

Arguments

data

List of data frames pertaining to a simulated study which is passed through in a list of snapshots

previous_data

List of data frames pertaining to prior snpashot with relation to data

spec

A list representing the combined specifications from CombineSpecs across the desired domains for a simulation.

...

Additional arguments to be passed.

Value

a data.frame pertaining to the raw dataset plugged into STUDCOMP.yaml

See also